reorder tests to provide a complex/offline/online distinction, also speeding up distributed test runs.
